lib: refactor new mixer out of 033_channel_strip
