+++ src/clover_sketches_lib/lib/mixer/draw_mixer_033.hpp
@@ -4,7 +4,7 @@
 // Copyright (C) 2025  Rob W. Albus
 // Licensed under the GPLv3. See LICENSE for details.
 
-#include "mixer.hpp"
+#include "mixer_033.hpp"
 
 // void mixer_component(const char* id, composition_mix* mixer);
 void draw_mixer_033(const char* id, mixer_033& mixer);
