+++ src/sketches/033_channel_strip/graph/patches/patches.hpp
@@ -9,10 +9,9 @@
 #include "lib/fdn/fdn8_023.hpp"
 #include "lib/fdn/fdn8_stereo_030.hpp"
 #include "lib/kick_drum/kick_drum.hpp"
+#include "lib/mixer/mixer_033.hpp"
 #include "lib/peq/peq.hpp"
 #include "lib/subtractive_synth/subtractive_synth.hpp"
-
-#include "graph/instrument/mixer.hpp"
 
 struct patch_drums_t {
     patch_drums_t();
